# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          }
        },
        "summary": "用户注册",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "用户登录",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/oauth/wechat-web/authorize": {
      "post": {
        "operationId": "AuthController_createWechatWebAuthorizeUrl_v1",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthAuthorizeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAuthorizeResponseDto"
                }
              }
            }
          }
        },
        "summary": "创建微信网页登录授权地址",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/oauth/wechat-web/callback": {
      "post": {
        "operationId": "AuthController_loginWithWechatWeb_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthCallbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "微信网页登录回调登录",
        "tags": ["Auth"]
      },
      "get": {
        "operationId": "AuthController_redirectWechatWebCallback_v1",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "description": "OAuth 授权码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "description": "授权时生成的 state",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to desktop callback URI"
          }
        },
        "summary": "微信网页登录浏览器回跳",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/oauth/wechat-mobile/callback": {
      "post": {
        "operationId": "AuthController_loginWithWechatMobile_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthCodeCallbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "微信移动端登录回调",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "用户登出",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseDto"
                }
              }
            }
          }
        },
        "summary": "刷新令牌",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/send-verification-code": {
      "post": {
        "operationId": "AuthController_sendVerificationCode_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendVerificationCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendVerificationCodeResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "验证码接口请求过多"
          }
        },
        "summary": "发送邮箱验证码",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailResponseDto"
                }
              }
            }
          }
        },
        "summary": "验证邮箱",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "验证码接口请求过多"
          }
        },
        "summary": "忘记密码",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "summary": "重置密码",
        "tags": ["Auth"]
      }
    },
...
    "/api/v1/user/data-export-requests": {
      "post": {
        "operationId": "DataExportController_createRequest_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDataExportRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataExportRequestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new data export request",
        "tags": ["Data Export"]
      }
    },
    "/api/v1/user/data-export-requests/latest": {
      "get": {
        "operationId": "DataExportController_getLatestRequest_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataExportLatestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get the latest data export request",
        "tags": ["Data Export"]
      }
    }
  },
  "info": {
    "title": "Lucent API",
    "description": "Lucent 后端 API 文档",
    "version": "1.0"
  },
  "components": {
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱地址",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "密码（8-32位，需包含大小写字母和数字）",
            "example": "Passw0rd123",
            "minLength": 8,
            "maxLength": 32
          },
          "code": {
            "type": "string",
            "description": "邮箱验证码",
            "example": "123456"
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "example": "小明",
            "maxLength": 20
          }
        },
        "required": ["email", "password", "code"]
      },
...
      "TokensDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "访问令牌"
          },
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌"
          },
          "expiresIn": {
            "type": "number",
            "description": "访问令牌过期时间（秒）",
            "example": 3600
          }
        },
        "required": ["accessToken", "refreshToken", "expiresIn"]
      },
      "RegisterDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserBriefDto"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokensDto"
          }
        },
        "required": ["user", "tokens"]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "结果码",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示消息",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/RegisterDataDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱地址",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "密码（与验证码二选一）",
            "example": "Passw0rd123",
            "minLength": 8,
            "maxLength": 32
          },
          "code": {
            "type": "string",
            "description": "邮箱验证码（与密码二选一）",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": ["email"]
      },
...
      "LoginDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserFullDto"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokensDto"
          }
        },
        "required": ["user", "tokens"]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "结果码",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示消息",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/LoginDataDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "OAuthAuthorizeDto": {
        "type": "object",
        "properties": {
          "callbackUri": {
            "type": "string",
            "description": "授权完成后的客户端回跳地址。桌面端支持 loopback 地址，Web 端支持可信 CORS origin 下的 /login/oauth/wechat。",
            "example": "http://127.0.0.1:49152/oauth/wechat"
          }
        }
      },
      "OAuthAuthorizeDataDto": {
        "type": "object",
        "properties": {
          "authorizeUrl": {
            "type": "string",
            "description": "第三方授权地址"
          },
          "state": {
            "type": "string",
            "description": "本次授权 state"
          },
          "expiresIn": {
            "type": "number",
            "description": "state 过期时间（秒）",
            "example": 600
          },
          "callbackUri": {
            "type": "string",
            "description": "客户端回跳地址。桌面端 loopback 或可信 Web 回调登录时返回。",
            "example": "http://127.0.0.1:49152/oauth/wechat"
          }
        },
        "required": ["authorizeUrl", "state", "expiresIn"]
      },
      "OAuthAuthorizeResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "结果码",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示消息",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/OAuthAuthorizeDataDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "OAuthCallbackDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "OAuth 授权码"
          },
          "state": {
            "type": "string",
            "description": "授权时生成的 state"
          }
        },
        "required": ["code", "state"]
      },
      "OAuthCodeCallbackDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "OAuth 授权码"
          }
        },
        "required": ["code"]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌"
          }
        },
        "required": ["refreshToken"]
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "结果码",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示消息",
            "example": ""
          },
          "data": {
            "type": "object",
            "description": "数据",
            "nullable": true,
            "example": null
          }
        },
        "required": ["code", "message", "data"]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌"
          }
        },
        "required": ["refreshToken"]
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "结果码",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示消息",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/TokensDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "SendVerificationCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱地址",
            "example": "user@example.com"
          },
          "scene": {
            "type": "string",
            "description": "验证码场景",
            "enum": ["register", "login", "reset-password", "change-email"],
            "example": "register"
          }
        },
        "required": ["email", "scene"]
      },
...
      "SendVerificationCodeResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "结果码",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示消息",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/CooldownMessageDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱地址",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "description": "验证码",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": ["email", "code"]
      },
...
      "VerifyEmailResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "结果码",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示消息",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/VerifyEmailDataDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱地址",
            "example": "user@example.com"
          }
        },
        "required": ["email"]
      },
      "ForgotPasswordResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "结果码",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示消息",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/CooldownMessageDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱地址",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "description": "验证码",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          },
          "password": {
            "type": "string",
            "description": "新密码（8-32位，需包含大小写字母和数字）",
            "example": "NewPassw0rd",
            "minLength": 8,
            "maxLength": 32
          }
        },
        "required": ["email", "code", "password"]
      },
...
      "CreateDataExportRequestDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": ["hospital", "monthly", "print"],
            "default": "hospital",
            "description": "Requested export kind."
          },
          "format": {
            "type": "string",
            "enum": ["pdf"],
            "default": "pdf",
            "description": "Requested export format."
          },
          "range": {
            "type": "string",
            "enum": ["last_7_days", "last_30_days"],
            "default": "last_7_days",
            "description": "Requested report range."
          }
        }
      },
...
      "DataExportRequestResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "Result code.",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "Message.",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/DataExportRequestDataDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "DataExportLatestResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "Result code.",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "Message.",
            "example": ""
          },
          "data": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DataExportRequestDataDto"
              }
            ]
          }
        },
        "required": ["code", "message", "data"]
      }
    }
  }
}